LayoutTests/storage/domstorage/sessionstorage/delete-removal.html

<html>
<head>
<script src="resources/clearSessionStorage.js"></script>
<script>

if (window.layoutTestController)
    layoutTestController.dumpAsText();

function log(a)
{
    document.getElementById('logger').appendChild(document.createTextNode(a));
    document.getElementById('logger').appendChild(document.createElement("br"));
}

function runTest()
{
    if (!window.sessionStorage) {
        log("window.sessionStorage DOES NOT exist");
        return;
    }
    
    log("foo is: " + sessionStorage.foo);
    sessionStorage.foo = "bar";
    log("foo is: " + sessionStorage.foo);
    delete sessionStorage.foo;
    log("foo is: " + sessionStorage.foo);
    sessionStorage["foo"] = "bar";
    log("foo is: " + sessionStorage.foo);
    delete sessionStorage.foo;
    log("foo is: " + sessionStorage.foo);
    sessionStorage.setItem("foo", "bar");
    log("foo is: " + sessionStorage.foo);
    delete sessionStorage.foo;
    log("foo is: " + sessionStorage.foo);
}

</script>
</head>
<body onload="runTest();">
This test makes sure that using the syntax `delete sessionStorage.keyName` works as an alias to `sessionStorage.removeItem(keyName).`<br><hr>
<div id="logger"></div>
</body>
</html>
