LayoutTests/storage/domstorage/sessionstorage/delete-removal-expected.txt

This test makes sure that using the syntax `delete sessionStorage.keyName` works as an alias to `sessionStorage.removeItem(keyName).`
foo is: undefined
foo is: bar
foo is: undefined
foo is: bar
foo is: undefined
foo is: bar
foo is: undefined
